feat: criacao do CharacterController e método de contabilizar métrica de cliques
